Move lambda tabs to separate components
